re #16444 Add test+logic to updateAppraisal
The Appraisals.updateAppraisal method takes as a parameter an appraisal
pojo. It then uses that value and tries update it. It saves the
associated assessments. IF the goals have been modified, it creates a new
log for that specific goal. Added tests for this method.

I tried to use a whitelist of fields to only save a subset of fields, but
that was a no go. Hibernate does not support a whitelist when saving
objects. Also if some of the values are null, we risk loosing data.