re #16444 Add db dtd and dbunit dataset
The database dtd is used by dbunit (similar to rails fixtures) to
specify what are allowed values in table columns. The dataset is an
xml that stores an xml representation of db table rows. The one sad
thing is that when exporting the dataset it is not very smart about
foreign key constraints so rows are not inserted in the correct
order.