re #16444 Add DBUnit.java to load seed test data
DBUnit.java takes care of reading the hibernate xml configs to get
the username and password. It uses this to connect to the test db and
load test data. In order to make parsing the xml configuration files
used by Hibernate a bit easier, I am using the apache commons
configuration and logging jars.

Also changed some of the HibernateUtil method variables to constants to
make it easier to store in one place the name and path to the dev and
test hibernate xml configuration files.