re #16444 Remove Business hbm + pojo
We are not going to use a BusinessCenter pojo or hbm file anymore.
Instead any tables that need to store the business center, will use a
String. The business center table will still be there, but only contain
a string for the name used for data integrity checks.

This commit also includes an update to the database.dtd and the full.xml
database seed file.