re #16444 Cleanup: pojo int to Integer and move resource bundle
Moving resource bundle from src.content to edu.osu.cws.pass.portlet.
Also the Appraisal had a few fields that were int, but they need to
be null since they are required. Changing things to Integer because the
primitive int datatype cannot be set to null. The supervisor in the jobs
table also points to a job instead of the employee. Adding a couple more
rows the the full.xml dataseed.