re #16444 Add getAllMyActiveAppraisals + tests
Adding getAllMyActiveAppraisals method to the model class. The first
method is public and it gets a session and then other one is a private
method that does all the work and gets as parameter the hibernate
session. Also added a test to check make sure the correct set of
appraisals are returned containing the right entries.