re #16444 Add portlet init method
Adding the portlet's init method that basically calls various hibernate
helper models to get list of admins, permissions, appraisal steps and
stores the maps in the portlet context. Also added a method called
stackTraceString that takes in a exception and returns a string of the
stack trace. This is needed so that we can log things to the luminis
log.