re #16444 Add getRole & isSupervisor method + test
The Jobs.isSupervisor method takes in a job and pidm. It traverses up
the supervising chain of that job and checks to see if one of the
supervisors matches the given pidm. The getRole method takes in a job
and a pidm. It checks to see whether the pidm matches the employee,
supervisor, reviewer or upper supervisor.