re #16444 Add appraisal & assessment dynamic-insert/update
Adding the property dynamic-insert and dynamic-update set to true for
the Appraisal and Assessment Hibernate bean classes. This is specified
in the hbm.xml files. I also split a couple of the many-to-one mappings
in those files so the lines wouldn't be too long.

The whole point of adding dynamic update/insert was to help the
updateAppraisal in the Appraisals class only update/insert the necessary
columns. I updated that method as well as test methods to reflect this.
Basically the session and transaction are now initialized outside of the
updateAppraisal method.