re #16444 Add properties file + refactor code
The file default.properties is loaded first and then it tries to load
docroot/WEB-INF/src/hostname.properties. These files get loaded only
once during the first time the portlet gets a request. The hostname in
the name of the file is needed so that the portlet can uniquely
identify which properties file to load. Right now the only property
that is stored in this file is which hibernate configuration file to
load. The hibernate configuration file includes credentials to login
to the db and connection settings.

Because loading the properties file required knowing the hostname of the
portal server, I had to move the code that was in the init method to a
separate portletSetup(request) method, which loads the initial
variables that don't change much and are related to permissions. This
method gets called
every time a doView, processAction or serveResource request happens.

The addition of the properties file caused a refactoring of the testing
code.  By default HibernateUitl, now loads the test hibernate config
file. The tests don't have to specify which hibernate config file to
load. The setEnvironment method in HibernateUtil was renamed to
setConfig which specifies what hibernate config file to load.

Adding *.properties to ignore list