re #16444 Update db fields related to appraisals
Updating the hibernate mapping files, java pojo, util classes, test
class and jsp files for appraisal, assessment, and goal_logs. The
updateAppraisal method in the Appraisals uitl class was updated as
well. The goal_logs includes a new field: type, which is used to figure
out if the goal entered is for the first set of goals or appended goals

Also,