re #16444 Fix bug that displayed my team section all the time
Added jobs.isSupervisor(int pidm) method that can check whether user
is a supervisor. This also includes tests to check the method works as
expected along with some extra test seed data.

The myTeam.jsp now checks that the user is a supervisor before displaying
that section. Also the actions class checks that the user is a supervisor
before adding required actions for supervisor.