re #16845 Remove skipDoView flag from portlet class
The skipDoView flag was used to let the doView method whether or not it
needed to call the delegate method. The skipDoView was set to true by
the processAction method and the actions' class methods. We can
accomplish the same by using the viewJSP property in the portlet class
and checking to make sure it's not null. If viewJSP is null in the
doView method, then we call the delegate method.