re #16845 Refactor error handling
Refactor the manner in which error handling is performed. Added an jsp
view: error.jsp which is used whenever an error is encountered. Also a
class edu.osu.cws.util.ExceptionHandler was added to handle exceptions
in the application. Made sure that the various util classe are throwing
errors and the actions class do not catch them. The delegate method in
the PASSPortlet class takes care of handling most of the exceptions by
instanating the ExceptionHandler. The portletSetup method also throws
errors instead of trying to catch them and the doView and processAction
take care of handling that error. Updated javadocs to match the errors
that were being thrown by methods