re #16845 Update view data: mutable and table names
Set mutable=false for both of the views in the hbm files. This should
tell Hibernate to prevent any changes for those objects. Also the
test db will be using tables and views for the job and employee data.
Thus, this required renaming the table names in the seed data file.