re #16898 Jobs.positionClass == pcls_code
The positionClass column in the jobs table was the same as the
pcls_code. Adding the column attribute to the  positionClass property
in the hbm file to specify the name of the column. Removing the old
pclsCode property from the job hbm file. Updating the dataseed file
to remove the positionClass column from the test db data.