re #16898 Update logic to fetch supervisor
Before the logic and db used to allow a job to have no employee
associated with it. Now, the job table uses a composite id which
contains the employee id. When we fetch the supervisor, we check
that the supervisor.status is 'A'ctive and that the supervsiro.employee
.active is 'A'ctive.
Also added some comments to the status properties in jobs and employees
models. Also refactored the supervisor methods in the JobMgr class to
use supervisorJob instead of current for some of the local variables to
improve readability.