re #16267 Criteria delete functionality
The delete criteria button in the evaluation criteria page now works.
Added some tests. If js is enabled, a pop-up is displayed and the
deletion is done using ajax. Otherwise, we display a new page to let
the user confirm the deletion. The various list criteria methods are
updated to only fetch the non-deleted criteria. The PASSPortlet class
now has some logic in the serveResource method to fetch calls to the
respective Actions class methods. Also added a helper method in the
Actions class called isAJAX to help determine whether or not the
request is ajax.