re #16267 Edit Critieria piece
Added Actions.editCriteria method which uses the criteria/add.jsp to
provide the edit form. The CriteriaMgr edit method is the one that
handles the edit logic. There are three possible scnearios: only area
is changed, only description or both. Each of these cases is handled.
Also if the option to propagate the changes is selected, the assessment
rows of all the non-closed and non-completed appraisals are updated.
Added some tests to check the three scenarios and propagate edits.
The user cannot change the appointment type in the edit form.