re #16890 Reviewer piece
Provides list, add and delete of reviewer users. Only master admin
users can add and delete reviewers. Validation is in place through js
to make sure that the onid username is not empty in the add form. There
is also validation in the java code to check and make sure that the user
is not already a reviewer for that business center and that it's a valid
active employee. Deleting or adding a reviewer resets the reviewers list
stored in the portlet context. Had to add a pojo, mgr and hbm files for
business center so that we can render the drop down list of business
centers using data from the db.

Includes tests to check: reviewers are listed correctly in the
expected order; reviewers can be deleted; reviewers who are not
active are not added; reviewers can be added; employees who are
reviewers for a different business center can be added as reviewers
for a new business center.