re #16890 Update criteria sequence
The js sorting of criteria is now in place. This is only supported via
js right now. The whole update criteria sequence also needs a bit more
thought during phase 2. We are updating the sequence of all the
involved evaluation criteria, but this also updates the sequence of
all the performance evaluations that have are using that evaluation
criteria. Add tests for updateSequence method in criteria mgr class.