re #16940 UI Notification refactoring
Several refactoring items related to notification:
1) The AppraisalMgr methods that grabbed the employees and team
appraisals used to return a map with a limited set of fields. Now they
return an Appraisal pojo with the few fields populated
2) The notification piece is split into my evaluation actions and my
administrative actions
3) The resource bundle keys for the actions required have been updated
along with the RequiredAction.setAnchorText method which now takes a
look at a configuration object to determine the # of days when an item
is due.
4) The notification piece is now displayed in a collapsible section
5) Renamed ConfigurationMgr2 -> ConfigurationMgr
6) Added CSS modifications
7) Updated status keys in the resource bundle file