re #16898 Fix performance issue in review list and search
Whenever the user was viewing a list of pending reviews, the load
time was really bad. It would take between 9-13 seconds for the page
to load. After doing some debugging, it was the sql query that was
taking a really long time. Fetching the supervisor information for
the pending appraisals was taking too long. I removed it from the
review list and appraisal search page for now. I had to modify
the hql queries and the appraisal constructor used by these.