re #16824 Refactor request.setAttribute in Actions class
Before we were using request.setAttribute to pass objects to the
jsp views. This became problematic when we are dealing with
methods called by processAction because processAction provides the
actionRequest, which does not allow us to set attributes for the view.
Instead, we now let the action methods set the complex objects in a
map. Then the doView methods takes care of calling
setRequestAttributes(renderRequest) which sets the elements in the
map for the view jsp to access.

This bug was identified when we tested the portlet in luminis 5.0.2