re #17676 Add Tom's query to fetch the onid username
Luminis doesn't store the onid username in the user.login.id or
user.name.nickName which is the standard for jsr 286. Instead
starting with LP5.0.1 they now store a random number. We are using
a query that Tom (ECS dba) provided us to fetch the onid username
of a user providing the screenName.