re #17821 get displayAppraisal to work with lazy loading
Getting displayAppraisal to work with lazy loading. The
displayAppraisal method in the Actions class initializes the necessary
collections before passing the appraisal pojo to the request. The main
change is that the appraisal status viewable by end users is now
calcualted by Appraisal.getViewStatus. Before we were setting the
status to display to the user in the status property and this was
causing lazy loading to save the status back in the appraisal.