Add new field to graylog messages for environment
re #18208

In order to make tracking error messages in graylog a bit easier, we
are now adding a new field to the graylog messages. I also removed
the clientHost parameter from the properties file and logger
constructor because they were not used. The logger class uses a java
method to figure out the real hostname of the server.