Use wrapper methods to access Session attributes
re #18333

Some of the values that we were trying to access form session
such as the review list, my active appraisals and my team's appraisals
were null. Now, we are using wrapper methods that check these
values in session first for null and if they are null then we get
them from the db and update the value in session.