Refactor js confirm pop-up
re #18494

The confirm dialog box wasn't display the text of the button for the
links in the action menu bar. Refactor the code so that we use js
to figure out the alt title of the image or the value of the
anchor when the user clicks on a menu item.