Allow smaller data points to group into "other"
re #19326

If the amount of data to be displayed in the chart, we grouped the lower
data points into a new category called "other". This maximum # of data
points allowed in the charts is controlled by a configuration option.
We basically grab the chart data and we make a copy of it. This copy of
the data has the "other" applied to it. In the js code, we have two
data tables, one for the chart (trimmedChartData) and the original
chartData used to display the data table at the bottom.