Fix bug in chart grouping other
re #19326

The grouping "other" was showing up in the data table listed below
the chart. The problem was that when trimming the data, we were using
sublist and adding to that the "other" group. The problem was that
subList didn't return a new list, instead it was returning a pointer
to the existing list.