Fix how appraisal.overdue is calculated
re #19003

Before we were leaving appraisal.overdue as it is if it didn't need to
be updated. This was creating the reports piece to report extra records
as being overdue when they weren't. We now set the overdue value to -999
if the record is not overdue.
Also refactored the newStatus code logic.