Allow for null values in Appraisal constructor
re #19003

The Appraisal constructor used by the reports piece had int as the
data type for the overdue property. This caused an error. Now, the
data type is set to Integer, to allow for nulls. If the value is null,
we set it to -999, which is the value we expect.