Fix permission issue w/ goal/results
re #21578

Fixing an issue where sometimes the user couldn't view the goals and
employee results of an appraisal. The code caches in the portlet
context the permission rules. When the appraisal was fetching and
modifying the permission rule it was the same permission rule object
used for all appraisals of that same status and role.