Include the due date in reminder emails
re #18683

To add the Due date into the email, we have to write a method call getDueDate in mailer, we will get the status of current appraisal.
If the status is "--Due", we can use EvalsUtil.getDueDate to use this status as the keywords to find the related "betweenDays"
 from configration and calculate the Duedate.
If the status is "--OverDue" we have to replace the "overDue" into "Due' and use EvalsUtil.gerDueDate to get the due date because
 configration does't contain the overDue status.
Then we have to update each related emai lbody method as shown in the list above.