Initial work on adding new assessments
re #23483

Adding new assessments to the form now works. It uses js to clone the
first assessment in the form, clears the data and sets the html ids,
names and class elements.
The AppraisalsAction class was updated to handle the new assessments
being added. The code still needs to be refactored, but I will handle
that in the next commit.
Deleting newly added assessments doesn't work. That will be fixed in the
next commit.