Recursively copy trial associations in appraisal
re #23499

When we create a first appraisal after annual, we call the
createFirstAnnual method in Appraisal.java. This calls methods
that recursively copy over assocations from trial
(copyPropertiesFromTrial).

The hbm.xml files were modified to specify cascade in the assocations.
This allows us to call save() on the Appraisal pojo and it takes care
of saving the other associations.