Empty goals are not saved
re #23775

If the user has empty goals, they will not be saved by the form.
JS handles the deleting of empty goals so they are not saved in the db.
I had to move the goals validation to be within the same js block
used by the previous js code so that it can call the assessment delete
function.

We delete an assessment via js rather than ignore saving them in the
java code because in the java code, we get a list of form fields
submitted. They are not structured. In js, we can easily target
the empty assessments and delete things with just 1 function call.