Add salary recommendation js logic
re #24444

The recommended increase is displayed or set based on the selected
rating. The salary after increase is calculated. Validation was added
for recommended increase. Resetting the value or html properties of
salary increase.

The salary object is saved in the java side. We also perform validation
on the java side as well.