Classified IT salary validation & range changes
re #24740

When rating is 3, the recommended increase is 0. When rating is 2, it
is a fixed value. When rating is 1 it is a range and the user can enter
a value. The fixed value and range for rating 1 & 2 change depending
on whether the current salary is above or below the control point.