Fix save draft null pointer error
re #24447

AppraisalsAction.getAppraisalStep() didn't use to return null. Thus the
calling code wasn't checking for nulls. In a previous commit, we added
the "Default" appraisal steps. This new piece of code was returning null
and the calling code was throwing a NPE. The calling code now checks for
nulls values being returned by this method.