Add more flexibility to goal validation
re #24823

Before the goal validation was strict and it would not allow the
form to validate if there was at least one empty goal with a criteria
checked. Now, we ignore the goals with empty text and checked criteria.
We assume a checked criteria with empty goals to be a user error and
ignore them.