Fix # of days in goals reactivation required action
re #25635

The goals reactivation request due was using the incorrect # of days.
This was caused because the appraisal objects loaded during login time
don't contain the goal version associated to them. This was causing the
default reference date of appraisal.start_date to be used when
calculating the due date for goal_reactivation request.