Fix error in creation of assessments
re #25638

When a new assessment is created from goals reactivation, it was using
the current criteria list from the db. The current list of criteria
areas in our tests was different than the initial criteria that were
available when the appraisal was created.

Now, when a new assessment is created when goals are reactivated it
uses the criteria list that are associated with the assessments with
the first approved goal version.