The js code was expecting the button to be clicked
re #25819

Since the problem happened when the user used the enter key to
submit the form, the validation js was not being fired up correctly.
There is now an extra class in the form element based on the evaluation
status. We use this in the appraisalDue status to check form
validation.