Fix error in pdf legend
re #25868

The pdf legend was trying to use the first goal version it found to
generate the legend. Since the goals versions aren't sorted, sometimes
it would find a goal version that was a denied reactivation. This
caused a NPE when it tried to access the assessments for the denied
goal version.