Unfinished bug fix for required action due date
re #25868

Some of the due date operations & comparisons use new DateTime().
Since we only show days, this caused an inconsistent bug. Depending
on the time of the day sometimes it would say something is due on 9
or 10 days. It was using the hours within the datetime objects to
calculate the number of full days before something was due. It was
rounding down.