Fix date in reactivation requested email
re #25908

The reactivation requested email was assuming that it was only sent out
right after the employee requested the reactivation. Now it asks the
db for the reactivated goal version. If the goal version reactivated
is null it means it hasn't been saved in the db and the date is now.
Otherwise we use the date from the db.