Remove logger dependency on action helper
re #26228

The logger object in portlet setup was created after db connection and
action helper creation. This was done to use a method from the action
helper to easily grab the evals properties. The logger object is now
created as soon as possible during portlet setup to avoid dependencies
on db and action helper.