Data loss if user fails validation
re #26421

The js code that saved the evaluation form using json was appending the
json data multiple times into the form. This caused there to be
multiple copies of the form data and only the first copy of the data
was being saved. This was causing data loss because a person can
submit the form empty first and get a validation error. After getting
the validation error message if the user submitted form data filled
out, this new data wasn't being saved.