The criteria checkboxes not always display in pdf
re #26524

If the # of criteria for a given goal is less than 4, the criteria
were not displayed at all in the form. This defect has probably been
in production for a while, but we didn't run into it. The column
span for the assessment criteria was set to 10, and there were a total
of 40 columns in the criteria tabl. If the # of columns used up by the
cells added to the table was less than 40, the row was not displayed
at all.

The column span calculation for the evaluation criteria now takes into
account the # of assessment criteria.