Fix performance issue with onid username query
re #26533

The database link luminis.osudb.osu is only needed when using OSDevl.
When this db link is used in production it causes some performance
issues. The query to get the onid username takes about 7-8 seconds.
The "@" is now part of the db link evals property value. Thus for
production and the default value of this property is blank. In ECS dev
environment, we use this db link since it's needed for devl and osdev2
doesn't have performance issues.