Autosave evals form
re #23787

Initial code for autosave via ajax of evaluation form. I would have
preferred to use a jquery plugin, but they either required a newer
version of jquery or they didn't have a license.

There are a couple of debug messages in this commit. The number of
seconds before each save are hard coded and need to be moved to the
configurations table. The UI for autosave functionality also needs to
be thought out a little bit more.