Add accessible messages - add/delete assessment
re #23794

When EvalS adds/deletes goals from the evaluation form it uses JS.
We now use a hidden aria-live div to tell the screen readers about
the add/delete of goals. This is something that visual users find
obvious. The css to place the aria-live div outside of the page is
part of the theme.