Fix Job equals comparison
EV-105

The Job.equals method was using the id value to compare if two objects
were the same. We used only the id since sometimes we don't load all the
properties of the Job object. The id is not a valid comparison since it
is a composite id, so the id is set to 0 for all objects.