Use the right appraisalStep to send mail
EV-127

Even though the correct status is used to update the evaluation record,
the email code was using the email_type from the appraisal_steps table.
The code now checks if the status used to update the evaluations is
different than the one in the appraisal steps, if it is, it uses the
correct email_type.