Send overdue emails if no due emails were sent
EV-144

The frequencyConfig with a value of -1 was preventing Overdue emails
from being sent. There are times, when no due email gets sent because
the evaluation is late. In order to notify those employees once, the
code now overwrites the frequencyConfig if no due and no overdue
emails have been sent.