Fix bug preventing deleting of goals
EV-514

When goals are deleted and the evaluation goals are submitted, it
was causing a js error. The jQuery filter used to find the deleted
hidden input was too broad for the newer version of jQuery. Modified
the find filter to specify that the hidden elements are supposed to be
of input hidden.