Changed the logic behind the isOverdue() function
EV-519

Previously, the isOverdue() function would check to
see if the isDue() returned <= 0. Only a return value of
<= 0 is attributed to the status of "overdue". However, the
function could have been checking if isDue() == 0 because of
the time of day that the cron job that updates appraisals
and such occur. It would make sense if it occurred later
in the day, rather than the start of the day.
In addition, I noticed that the signatureDue message for
signatureDue-today was not quite right, so I changed that
as well.