Added test for addAssessment()
EV-518

Test requires transaction as an error will be thrown
without one. I believe that this might be causing a problem
with multiple test runs. Currently, the test assumes that nothing
it does is permanently saved, so the next time it is called, it
should have the same results. However, I have not found this to
be the case. It only works if I restart my MAMP servers between
each test run.