Error prevents supervisor from submit appraisal
EV-606

The method that checks the next valid status was converting employee
review due to "employee review overdue". The employee review status
becomes expired after the status. Because "overdue" was not a valid
state for the employee review stage, the item was missing from the
Language.properties file. This caused the supervisor action of
submitting the appraisal to abort.